Fixed gallery page to delete image above button
After uploading a photo and hitting delete it would delete all photos on the page instead of the one you clicked "Delete" on. Now, it deletes the photo you clicked "Delete" on instead of deleting all images on the gallery page. I updated the recipelists.sql file to include a primary key for 'id' in the images table and added an auto increment to the 'id' field so that it deletes only the photo you want to remove from the "Gallery of Cooked Recipes" web page. Finally, I deleted an extra "</div>" in register.html (line 40) and showProfilePage.php (line 104), added an extra quotation marks to the end of class="search-textbox" and deleted </input> (line 28) in showAddForm.php, and added a closing </span> to showTopMenu.php (line 22).